Run weekly builds and force rebuilding everything
With this triggering a rebuild of psptoolchain-extra, it will make sure that everything is rebuild once a week. This is because of the cascade that is set up to go all the way to psp-libraries. I set it to 17:00 on Friday, because if it breaks I'll probably be fixing it on the weekend anyway.